Add post authguard

+++ src/posts/posts.service.ts
@@ -6,6 +6,7 @@ import createPostDto from './dto/create-post.dto';
 import editPostDto from './dto/edit-post.dto';
 import addCommentDto from './dto/add-comment.dto';
 import mongoose from 'mongoose';
+import { types } from 'util';
 
 // import createUserDto from './dto/create-user-dto';
 
@@ -18,10 +19,10 @@ export class PostsService {
         return await this.postModel.find({});
     }
 
-    async create(post: createPostDto): Promise<any> {
+    async create(post: createPostDto, userId: string): Promise<any> {
         const postEntity: Post = await this.postModel.create({
             ...post,
-            userId: mongoose.Types.ObjectId(), //fake userId
+            userId: Types.ObjectId(userId), //fake userId
             timestamp: new Date(),
             isEdited: false,
             comments: [],
@@ -30,9 +31,11 @@ export class PostsService {
         return postEntity;
     }
 
-    async editPost(post: editPostDto): Promise<any> {
-        //TODO: check role
-        return this.postModel.updateOne(
+    async editPost(post: editPostDto, userId: string): Promise<any> {
+        const res = await this.postModel.findOne({ _id: post.postId });
+        if (res.userId.toHexString() !== userId) throw new Error('auth err');
+
+        return await this.postModel.updateOne(
             { _id: post.postId },
             {
                 ...post,
@@ -42,8 +45,9 @@ export class PostsService {
         );
     }
 
-    async deletePost(postId: Types.ObjectId): Promise<any> {
-        //TODO: check role
+    async deletePost(postId: Types.ObjectId, userId: string): Promise<any> {
+        const res = await this.postModel.findOne({ _id: postId });
+        if (res.userId.toHexString() !== userId) throw new Error('auth err');
         return this.postModel.deleteOne({ _id: postId });
     }
